libpcp/ipc.c - tweak last fix
Use same (int) cast for printf %d that we use elsewhere.

+++ src/libpcp/src/ipc.c
@@ -196,7 +196,7 @@ __pmSetDataIPC(int fd, void *data)
     }
 
     if (pmDebug & DBG_TRACE_CONTEXT)
-	fprintf(stderr, "__pmSetDataIPC: fd=%d data=%p(sz=%d)\n", fd, data, ipcentrysize - sizeof(__pmIPC));
+	fprintf(stderr, "__pmSetDataIPC: fd=%d data=%p(sz=%d)\n", fd, data, (int)(ipcentrysize - sizeof(__pmIPC)));
 
     dest = ((char *)__pmIPCTablePtr(fd)) + sizeof(__pmIPC);
     memcpy(dest, data, ipcentrysize - sizeof(__pmIPC));
@@ -226,7 +226,7 @@ __pmDataIPC(int fd, void *data)
     }
     source = ((char *)__pmIPCTablePtr(fd)) + sizeof(__pmIPC);
     if (pmDebug & DBG_TRACE_CONTEXT)
-	fprintf(stderr, "__pmDataIPC: fd=%d, data=%p(sz=%d)\n", fd, source, ipcentrysize - sizeof(__pmIPC));
+	fprintf(stderr, "__pmDataIPC: fd=%d, data=%p(sz=%d)\n", fd, source, (int)(ipcentrysize - sizeof(__pmIPC)));
     memcpy(data, source, ipcentrysize - sizeof(__pmIPC));
 
     PM_UNLOCK(__pmLock_libpcp);
